FIX: Simpler makefile FIX: Size_t instead of Int FIX: Norminette Part I / Part II ADD: Many bonus fonction

+++ ft_putendl_fd.c
@@ -6,7 +6,7 @@
 /*   By: dernst <dernst@student.42lyon.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/09 18:47:23 by dernst            #+#    #+#             */
-/*   Updated: 2024/11/11 17:08:24 by dernst           ###   ########lyon.fr   */
+/*   Updated: 2024/11/12 00:43:15 by dernst           ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_putendl_fd(char *s, int fd)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
