+++ ft_strjoin.c
@@ -6,7 +6,7 @@
 /*   By: dernst <dernst@student.42lyon.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/26 15:06:52 by ernstdevan        #+#    #+#             */
-/*   Updated: 2024/11/08 11:42:15 by dernst           ###   ########lyon.fr   */
+/*   Updated: 2024/11/12 00:49:51 by dernst           ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 char	*ft_strjoin(char const *s1, char const *s2)
 {
 	char	*str;
-	int		i;
-	int		j;
+	size_t	i;
+	size_t	j;
 
 	str = malloc((ft_strlen(s1) + ft_strlen(s2)) + 1);
 	j = 0;
